[measure] move to sub-folder
