+++ c++/measures/average_sign.cpp
@@ -18,18 +18,27 @@
  * TRIQS. If not, see <http://www.gnu.org/licenses/>.
  *
  ******************************************************************************/
-#pragma once
-#include "./qmc_data.hpp"
+#include "./average_sign.hpp"
 
 namespace cthyb {
 
-struct measure_density_matrix {
- qmc_data const& data;
- std::vector<matrix_t>& block_dm; // density matrix of each block
- mc_weight_t z = 0;
+measure_average_sign::measure_average_sign(qmc_data const& data, mc_weight_t& average_sign)
+   : data(data), average_sign(average_sign) {
+ average_sign = 1.0;
+ z = 0;
+ sign = 0;
+}
 
- measure_density_matrix(qmc_data const& data, std::vector<matrix_t> & density_matrix);
- void accumulate(mc_weight_t s);
- void collect_results(triqs::mpi::communicator const& c);
-};
+void measure_average_sign::accumulate(mc_weight_t s) {
+
+ sign += s * data.atomic_reweighting;
+ z += std::abs(data.atomic_reweighting);
+}
+
+void measure_average_sign::collect_results(triqs::mpi::communicator const& c) {
+
+ z = mpi_all_reduce(z, c);
+ sign = mpi_all_reduce(sign, c);
+ average_sign = sign / z;
+}
 }
